app/requirement.txt
connexion[swagger-ui]
python-consul
tornado
